requirements.txt
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.5.0
docstring-inheritance==1.1.1
fonttools==4.39.4
fsspec==2023.5.0
future==0.18.3
importlib-metadata==6.6.0
iso8601==1.1.0
kiwisolver==1.4.4
locket==1.0.0
matplotlib==3.7.1
ndtiff==2.0.0
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
partd==1.4.0
Pillow==9.5.0
psutil==5.9.5
pycromanager==0.27.5
PyLabSAS==2.0.3
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyqtgraph==0.12.4
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
scipy==1.10.1
six==1.16.0
toolz==0.12.0
zipp==3.15.0
